orginize structure
